github_agent/pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gitHub_a2a_agent"
version = "0.1.0"
authors = [
    { name = "Arnaud Legendre", email = "legendre@google.com" }
]
requires-python = ">=3.10"

dependencies = [
  "google-adk[a2a]>=1.31.0,<2.0.0",
  "pyarrow<24",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2,<7.0.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-dotenv", "python-dotenv"]

[tool.setuptools.packages]
find = {include = ["github_agent*"]}

[tool.setuptools.package-data]
github_agent = ["prompts/*.yaml"]
